# [BugFix] Optimize partition compensate strategy for performance(Part1)