--- a/fe/fe-core/src/main/java/com/starrocks/sql/common/QueryDebugOptions.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/common/QueryDebugOptions.java
@@ -31,6 +31,9 @@ public class QueryDebugOptions {
     @SerializedName(value = "enableMVEagerUnionAllRewrite")
     private boolean enableMVEagerUnionAllRewrite = false;
 
+    @SerializedName(value = "enableQueryTraceLog")
+    private boolean enableQueryTraceLog = false;
+
     public QueryDebugOptions() {
         // To make unit test more stable, add retry times for refreshing materialized views.
         if (FeConstants.runningUnitTest) {
@@ -60,6 +63,14 @@ public class QueryDebugOptions {
 
     public void setEnableMVEagerUnionAllRewrite(boolean enableMVEagerUnionAllRewrite) {
         this.enableMVEagerUnionAllRewrite = enableMVEagerUnionAllRewrite;
+    }
+
+    public boolean isEnableQueryTraceLog() {
+        return enableQueryTraceLog;
+    }
+
+    public void setEnableQueryTraceLog(boolean enableQueryTraceLog) {
+        this.enableQueryTraceLog = enableQueryTraceLog;
     }
 
     public static QueryDebugOptions getInstance() {