--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MaterializationContext.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MaterializationContext.java
@@ -15,15 +15,20 @@
 package com.starrocks.sql.optimizer;
 
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.starrocks.catalog.MaterializedView;
 import com.starrocks.catalog.Table;
+import com.starrocks.common.Pair;
+import com.starrocks.qe.SessionVariable;
 import com.starrocks.sql.optimizer.base.ColumnRefFactory;
 import com.starrocks.sql.optimizer.operator.OperatorType;
 import com.starrocks.sql.optimizer.operator.logical.LogicalAggregationOperator;
 import com.starrocks.sql.optimizer.operator.logical.LogicalOlapScanOperator;
+import com.starrocks.sql.optimizer.operator.logical.LogicalScanOperator;
 import com.starrocks.sql.optimizer.operator.pattern.Pattern;
 import com.starrocks.sql.optimizer.operator.scalar.ColumnRefOperator;
+import com.starrocks.sql.optimizer.operator.scalar.ConstantOperator;
 import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.MaterializedViewRewriter;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.MvUtils;
@@ -35,6 +40,7 @@ import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -79,6 +85,18 @@ public class MaterializationContext {
     // in Optimizer Transformation phase but not be really used.
     private long mvUsedCount = 0;
 
+    //// Caches for the mv rewrite lifecycle
+
+    // Cache whether to need to compensate partition predicates or not, and it's not changed
+    // during one query, so it's safe to cache it and be used for each optimizer rule.
+    // But it is different for each materialized view, compensate partition predicate from the plan's
+    // `selectedPartitionIds`, and check `isNeedCompensatePartitionPredicate` to get more information.
+    private Optional<Boolean> isCompensatePartitionPredicateOpt = Optional.empty();
+    // Cache a mv's pruned partition predicates in the rewrite because it's not changed through the context.
+    private Optional<ScalarOperator> mvPrunedPartitionPredicateOpt = Optional.empty();
+    // Cache partition compensates predicates for each ScanNode and isCompensate pair.
+    private Map<Pair<LogicalScanOperator, Boolean>, List<ScalarOperator>> scanOpToPartitionCompensatePredicates;
+
     public MaterializationContext(OptimizerContext optimizerContext,
                                   MaterializedView mv,
                                   OptExpression mvExpression,
@@ -102,6 +120,7 @@ public class MaterializationContext {
         this.matchedGroups = Lists.newArrayList();
         this.mvPartialPartitionPredicate = mvPartialPartitionPredicate;
         this.refTableUpdatePartitionNames = refTableUpdatePartitionNames;
+        this.scanOpToPartitionCompensatePredicates = Maps.newHashMap();
     }
 
     public MaterializedView getMv() {
@@ -371,4 +390,67 @@ public class MaterializationContext {
         }
     }
 
+    /**
+     * What's the meaning of partition compensate in mv rewriting?
+     * <p>
+     * Because `PartitionPruner` will prune predicates associated with partition columns before mv rewrite,
+     * so we cannot use both current query/mv plan's predicates for rewrite, otherwise it may generate
+     * wrong rewrite results because of loss of pruned predicates.
+     * </p>
+     *
+     * <p>
+     *There are two choices to recover partition predicates from the current pruned plan:
+     *Method 1:
+     *  Because `PartitionPruner` keeps partitions to scan in `selectedPartitionIds` after complex pruning rules,
+     *  so use `selectedPartitionIds` to compensate complete partition ranges with lower and upper bound directly.
+     *
+     * Method 2:
+     *  `PartitionPruner` also will keep original predicates before partition pruning, so we can use original
+     *  pruned partition predicates as the compensated partition predicates directly.
+     *</p>
+     *
+     * <p>
+     * Method1(default) takes advantage of `PartitionPruner`'s prune abilities and deduces a new normalized partition
+     * predicate, but it may generate some redundant or noisy partition predicates to disturb rewriting.
+     *
+     * Method2(preferred) doesn't use the advantage of `PartitionPruner`'s prune abilities and treats query and mv's original
+     * queries to compare/rewrite, and it will simplify the rewrite routine.
+     * And also, MV rewrite's abilities should cover `PartitionPruner` and no use Method1 either.
+     * </p>
+     *
+     * <p>
+     * What's the current strategy to handle this?
+     *  Method1 is used by default to avoid breaking compatibility, but optimize it when MV is no need to compensate which
+     *  means MV's partitions can cover all needed partitions from Query.
+     * </p>
+     */
+    public boolean isCompensatePartitionPredicate(OptExpression queryExpression) {
+        if (!isCompensatePartitionPredicateOpt.isPresent()) {
+            SessionVariable sessionVariable = optimizerContext.getSessionVariable();
+            // only set this when `queryExpression` contains ref table, otherwise the cached value maybe dirty.
+            isCompensatePartitionPredicateOpt = sessionVariable.isEnableMaterializedViewRewritePartitionCompensate() ?
+                    MvUtils.isNeedCompensatePartitionPredicate(queryExpression, this) : Optional.of(false);
+        }
+        return isCompensatePartitionPredicateOpt.isPresent() ? isCompensatePartitionPredicateOpt.get() : true;
+    }
+
+    public boolean isCompensatePartitionPredicate() {
+        return isCompensatePartitionPredicateOpt.isPresent() ? isCompensatePartitionPredicateOpt.get() : true;
+    }
+
+    public ScalarOperator getMVPrunedPartitionPredicate() {
+        if (!mvPrunedPartitionPredicateOpt.isPresent()) {
+            List<ScalarOperator> mvPrunedPartitionPredicates = MvUtils.getMVPrunedPartitionPredicates(mv, mvExpression);
+            if (mvPrunedPartitionPredicates == null || mvPrunedPartitionPredicates.isEmpty()) {
+                mvPrunedPartitionPredicateOpt = Optional.of(ConstantOperator.TRUE);
+            } else {
+                mvPrunedPartitionPredicateOpt = Optional.of(Utils.compoundAnd(mvPrunedPartitionPredicates));
+            }
+        }
+        return mvPrunedPartitionPredicateOpt.get();
+    }
+
+    public Map<Pair<LogicalScanOperator, Boolean>, List<ScalarOperator>> getScanOpToPartitionCompensatePredicates() {
+        return scanOpToPartitionCompensatePredicates;
+    }
 }