--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MvRewriteContext.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MvRewriteContext.java
@@ -47,9 +47,6 @@ public class MvRewriteContext {
 
     private List<JoinDeriveContext> joinDeriveContexts;
 
-    // Whether to compensate partition predicate from the plan's `selectedPartitionIds`,
-    // check `isNeedCompensatePartitionPredicate` to get more information.
-    private final boolean isCompensatePartitionPredicate;
 
     public MvRewriteContext(
             MaterializationContext materializationContext,
@@ -58,8 +55,7 @@ public class MvRewriteContext {
             ReplaceColumnRefRewriter queryColumnRefRewriter,
             PredicateSplit queryPredicateSplit,
             List<ScalarOperator> onPredicates,
-            Rule rule,
-            boolean isCompensatePartitionPredicate) {
+            Rule rule) {
         this.materializationContext = materializationContext;
         this.queryTables = queryTables;
         this.queryExpression = queryExpression;
@@ -68,7 +64,6 @@ public class MvRewriteContext {
         this.onPredicates = onPredicates;
         this.rule = rule;
         this.joinDeriveContexts = Lists.newArrayList();
-        this.isCompensatePartitionPredicate = isCompensatePartitionPredicate;
     }
 
     public MaterializationContext getMaterializationContext() {
@@ -121,9 +116,5 @@ public class MvRewriteContext {
 
     public void setEnforcedNonExistedColumns(List<ColumnRefOperator> enforcedNonExistedColumns) {
         this.enforcedNonExistedColumns = enforcedNonExistedColumns;
-    }
-
-    public boolean isCompensatePartitionPredicate() {
-        return this.isCompensatePartitionPredicate;
     }
 }