--- a/fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewriteUnionTest.java
+++ b/fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewriteUnionTest.java
@@ -458,7 +458,7 @@ public class MvRewriteUnionTest extends MvRewriteTestBase {
                 "  |  predicates: 1: k1 > 1, 2: k2 LIKE 'a%'";
         String p =
                 "7:SELECT\n" +
-                "  |   predicates: 2: k2 LIKE 'a%', 1: k1 > 1";
+                "  |  predicates: 2: k2 LIKE 'a%', 1: k1 > 1";
         PlanTestBase.assertContainsIgnoreColRefs(p1, p);
     }
 }