… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MaterializedViewRewriter.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvUtils.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/rule/BaseMaterializedViewRewriteRule.java
… fe/fe-core/src/test/java/com/starrocks/planner/MaterializedViewLowCardTPCHTest.java
… fe/fe-core/src/test/java/com/starrocks/planner/MaterializedViewManualTest.java
… fe/fe-core/src/test/java/com/starrocks/planner/MaterializedViewTest.java
… fe/fe-core/src/test/java/com/starrocks/planner/MaterializedViewTestBase.java
… fe/fe-core/src/test/java/com/starrocks/schema/MSchema.java
… fe/fe-core/src/test/java/com/starrocks/schema/MTable.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MVPartitionCompensateOptBench.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRefreshAndRewriteIcebergTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewritePartitionTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewriteTestBase.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvUtilsTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/NormalizePredicateBench.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/PlanTestNoneDBBase.java
… fe/fe-core/src/test/java/com/starrocks/utframe/UtFrameUtils.java
… fe/fe-core/src/test/resources/sql/materialized-view/tpch-hive/q11.sql
… fe/fe-core/src/test/resources/sql/materialized-view/tpch/q5.sql