Why I'm doing:
When there are many materialized views, the mv rewrite optimizer time may cost too much. The main reason is that mv rewrite computes partition prune and cannonize predicates too much and there are a lot of repeat with them.


What I'm doing:

For the repeat compute, we can add some caches during the query's lifecycle. Add caches below:
- Add caches for per materialized view used for partition predicates mainly:
```
    // Cache whether to need to compensate partition predicates or not, and it's not changed
    // during one query, so it's safe to cache it and be used for each optimizer rule.
    // But it is different for each materialized view, compensate partition predicate from the plan's
    // `selectedPartitionIds`, and check `isNeedCompensatePartitionPredicate` to get more information.
    private Optional<Boolean> isCompensatePartitionPredicateOpt = Optional.empty();
    // Cache a mv's pruned partition predicates in the rewrite because it's not changed through the context.
    private Optional<ScalarOperator> mvPrunedPartitionPredicateOpt = Optional.empty();
    // Cache partition compensates predicates for each ScanNode and isCompensate pair.
    private Map<Pair<LogicalScanOperator, Boolean>, List<ScalarOperator>> scanOpToPartitionCompensatePredicates;
```
- Add caches for the query level no matter which materialized view: 

```
    // `mvQueryContextCache` is designed for a common cache which is used during the query's rewrite lifecycle, so can be
    // managed in a more unified mode. In the current situation, it's used in two ways:
    // 1. cache query predicates to its final predicate split per query because PredicateSplit's construct is expensive.
    // 2. cache query predicate to its canonized predicate to avoid one predicate's repeat canonized.
    // It can be be used for more situations later.
    private final Cache<Object, Object> mvQueryContextCache = Caffeine.newBuilder()
            .maximumSize(Config.mv_query_context_cache_max_size)
            .recordStats()
            .build();
```


I had two tests in my apple mac as below:
```
➜  ~ sysctl -a|grep cpu
kern.sched_rt_avoid_cpu0: 0
kern.cpu_checkin_interval: 4000
hw.ncpu: 8
hw.activecpu: 8
hw.perflevel0.physicalcpu: 6
hw.perflevel0.physicalcpu_max: 6
hw.perflevel0.logicalcpu: 6
hw.perflevel0.logicalcpu_max: 6
hw.perflevel0.cpusperl2: 3
hw.perflevel1.physicalcpu: 2
hw.perflevel1.physicalcpu_max: 2
hw.perflevel1.logicalcpu: 2
hw.perflevel1.logicalcpu_max: 2
hw.perflevel1.cpusperl2: 2
hw.physicalcpu: 8
hw.physicalcpu_max: 8
hw.logicalcpu: 8
hw.logicalcpu_max: 8
hw.cputype: 16777228
hw.cpusubtype: 2
hw.cpu64bit_capable: 1
hw.cpufamily: 458787763
hw.cpusubfamily: 4
machdep.cpu.cores_per_package: 8
machdep.cpu.core_count: 8
machdep.cpu.logical_per_package: 8
machdep.cpu.thread_count: 8
machdep.cpu.brand_string: Apple M1 Pro
```
### Test1

Run `testMVPartitionCompensatePerf1` 100 times with 10 materialized views:
```
    @Test
    @BenchmarkOptions(warmupRounds = 3, benchmarkRounds = 100)
    public void testMVPartitionCompensatePerf1() {
        testMVPartitionCompensatePerf(1);
    }
```

Before:
![image](https://github.com/StarRocks/starrocks/assets/5104975/001d4e1b-a9ca-492b-be12-ab965e4972f1)

After:
![image](https://github.com/StarRocks/starrocks/assets/5104975/c7759361-6724-49ff-ad95-78ee2a7840fe)

### Test2
Run `MVPartitionCompensateOptBench` all tests with 100 mv and benchmarkRounds=10:


Before | After(without mv limit 1000) | Default(with limit mvs)
-- | -- | --
3.16 | 1.27 | 1.13
7.49 | 1.81 | 1.23
4.13 | 1.44 | 1.03
2.68 | 1.06 | 0.95
2.64 | 1.05 | 0.95
2.71 | 1.05 | 0.9
4.89 | 1.44 | 1.02
2.45 | 1 | 0.86
2.48 | 0.95 | 0.89
2.26 | 0.95 | 0.88
2.13 | 0.92 | 0.83
```
   connectContext.getSessionVariable().setCboMaterializedViewRewriteCandidateLimit(1000);
   connectContext.getSessionVariable().setCboMaterializedViewRewriteRuleOutputLimit(1000);
```
Final Cache Stats:
Query2, one round:
```
PredicateSplitCache  Stats:CacheStats{hitCount=7863, missCount=16, loadSuccessCount=16, loadFailureCount=0, totalLoadTime=2402417, evictionCount=0, evictionWeight=0}, estimatedSize:16
, PredicateCanonizedCache Stats:CacheStats{hitCount=5273, missCount=50, loadSuccessCount=50, loadFailureCount=0, totalLoadTime=2449207, evictionCount=0, evictionWeight=0}, estimatedSize:50
```
Fixes https://github.com/StarRocks/StarRocksTest/issues/5178

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5